refactor : Service 모델 Dto로 변경
